+++ commands/tv/search.js
@@ -1,11 +1,16 @@
-const createTvShowItemModal = require('../../util/sonarr/createTvShowModal');
+const createTvShowItemModal = require('../../util/discord/createTvShowModal');
 const getTvShowInfo = require('../../util/omdb/getTvShow');
+const getTvdbIdByImdbId = require('../../util/thetvdb/getTvdbIdByImdbId');
 exports.run = (bot, msg, args = []) => {
   getTvShowInfo(args)
     .then((tvShow) => {
-      const e = createTvShowItemModal(tvShow);
-      e.setFooter(`Called by ${msg.author.username}`, msg.author.avatarURL);
-      msg.channel.send({ 'embed': e });
+      getTvdbIdByImdbId(msg.guild, tvShow.imdbID)
+        .then((res) => {
+          tvShow.tvdbID = res;          
+          const e = createTvShowItemModal(tvShow);    
+          e.setFooter(`Called by ${msg.author.username}`, msg.author.avatarURL);
+          msg.channel.send({ 'embed': e });      
+        });
     });
 };
 exports.conf = {
