+++ config/default.json
@@ -2,6 +2,10 @@
     "prefix": "!",
     "uuidv4": "",
     "omdb_apikey": "5af02350",
+    "thetvdb": {
+        "url": "",
+        "apikey": ""
+    },
     "tautulli": {
         "url": "",
         "apikey": ""
